[add] Add EnemyBar
敵のHPバーがHPに応じて赤くなるように変更

+++ Assets/Scripts/StageScene.cs
@@ -20,7 +20,7 @@ public class StageScene : MonoBehaviour
 
     //“G‚ÌHP‚ğİ’è
     [SerializeField]
-    public int EnemyHp;
+    public float EnemyHp;
 
     public GameObject Player;
